add api/v1 directoy and change user method to create
